+++ packages/vm/src/instance.rs
@@ -417,7 +417,7 @@ mod singlepass_test {
 
         let handle_used = gas_before_handle - instance.get_gas();
         println!("handle used: {}", handle_used);
-        assert_eq!(handle_used, 63554);
+        assert_eq!(handle_used, 63555);
     }
 
     #[test]
@@ -452,6 +452,6 @@ mod singlepass_test {
 
         let query_used = gas_before_query - instance.get_gas();
         println!("query used: {}", query_used);
-        assert_eq!(query_used, 23050);
+        assert_eq!(query_used, 23018);
     }
 }